% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{hc_inegi}
\alias{hc_inegi}
\title{Chart INEGI series}
\usage{
hc_inegi(hc, series, token, type = "line", names = NULL)
}
\arguments{
\item{hc}{highcharter object}

\item{series}{one or more INEGI series to chart. All will obey same type of chart. Must be same length and frequency.}

\item{token}{INEGI API token}

\item{type}{type of chart, to pass to highcharter}

\item{names}{if NULL will use metadata to name series, but can be overwriten with this parameter 

# @importFrom inegiR compact_inegi_series}
}
\description{
Charts one or more INEGI series using inegiR package.
}
\examples{
\dontrun{
library(higcharter)
library(inegiR)
highchart() \%>\% hc_inegi(c(s1, s2), token)
}
}

